template/template_stack: can't both be enforced yet conflicting (Closes #294)
The template and template_stack objects are enforced when running
plan/apply against imported objects; otherwise the engine would
want to completely replace the objects, which would likely cause
consistency issues with the rest of the configuration.

Since they are both enforced; they can't also be conflicting; that
leads to an impossible scenario where you must but can't define
both a template and template_stack.

Signed-off-by: Mathieu Trudel-Lapierre <mtrudel@stack8.com>